Removed Translation concept and replaced with Card Question and Card Answer system
Adjusted the project pretty heavily. This is because Google Translate API requires fees to use. So instead I've decided to simply make this a CardQuestion and CardAnswer based flash card instead of automatically translating the card. This might make things a lot easier in the future in case Google Translate incorrectly translates a phrase as well.